set product_supplierinfo_inter_company and stock_intercompany to rebel module
For product_supplierinfo_inter_company, this introduces a hard requirement to configure
a pricelist on each supplier. This makes the tests of other modules fail because they do
not set these.

For stock_intercompany I don't know why it was set to rebel, it was done before
but not updated in the copier template, so just updating copier to reflect reality.